Moves account settings into a sub view in settings.
